# Enable TensorPrimitive vectorization for Half for lots of methods