--- a/src/libraries/System.Numerics.Tensors/src/System/Numerics/Tensors/netcore/TensorPrimitives.BitwiseAnd.cs
+++ b/src/libraries/System.Numerics.Tensors/src/System/Numerics/Tensors/netcore/TensorPrimitives.BitwiseAnd.cs
@@ -23,7 +23,7 @@ namespace System.Numerics.Tensors
         public static void BitwiseAnd<T>(ReadOnlySpan<T> x, ReadOnlySpan<T> y, Span<T> destination)
             where T : IBitwiseOperators<T, T, T>
         {
-            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsShort<T, BitwiseAndOperator<short>>(x, y, destination))
+            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsInt16<T, BitwiseAndOperator<short>>(x, y, destination))
             {
                 return;
             }
@@ -45,7 +45,7 @@ namespace System.Numerics.Tensors
         public static void BitwiseAnd<T>(ReadOnlySpan<T> x, T y, Span<T> destination)
             where T : IBitwiseOperators<T, T, T>
         {
-            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsShort<T, BitwiseAndOperator<short>>(x, y, destination))
+            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsInt16<T, BitwiseAndOperator<short>>(x, y, destination))
             {
                 return;
             }