--- a/src/libraries/System.Numerics.Tensors/src/System/Numerics/Tensors/netcore/TensorPrimitives.BitwiseOr.cs
+++ b/src/libraries/System.Numerics.Tensors/src/System/Numerics/Tensors/netcore/TensorPrimitives.BitwiseOr.cs
@@ -21,8 +21,15 @@ namespace System.Numerics.Tensors
         /// </para>
         /// </remarks>
         public static void BitwiseOr<T>(ReadOnlySpan<T> x, ReadOnlySpan<T> y, Span<T> destination)
-            where T : IBitwiseOperators<T, T, T> =>
+            where T : IBitwiseOperators<T, T, T>
+        {
+            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsShort<T, BitwiseOrOperator<short>>(x, y, destination))
+            {
+                return;
+            }
+
             InvokeSpanSpanIntoSpan<T, BitwiseOrOperator<T>>(x, y, destination);
+        }
 
         /// <summary>Computes the element-wise bitwise OR of numbers in the specified tensors.</summary>
         /// <param name="x">The first tensor, represented as a span.</param>
@@ -36,8 +43,15 @@ namespace System.Numerics.Tensors
         /// </para>
         /// </remarks>
         public static void BitwiseOr<T>(ReadOnlySpan<T> x, T y, Span<T> destination)
-            where T : IBitwiseOperators<T, T, T> =>
+            where T : IBitwiseOperators<T, T, T>
+        {
+            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsShort<T, BitwiseOrOperator<short>>(x, y, destination))
+            {
+                return;
+            }
+
             InvokeSpanScalarIntoSpan<T, BitwiseOrOperator<T>>(x, y, destination);
+        }
 
         /// <summary>x | y</summary>
         private readonly struct BitwiseOrOperator<T> : IBinaryOperator<T> where T : IBitwiseOperators<T, T, T>