--- a/src/libraries/System.Numerics.Tensors/src/System/Numerics/Tensors/netcore/TensorPrimitives.BitwiseOr.cs
+++ b/src/libraries/System.Numerics.Tensors/src/System/Numerics/Tensors/netcore/TensorPrimitives.BitwiseOr.cs
@@ -23,7 +23,7 @@ namespace System.Numerics.Tensors
         public static void BitwiseOr<T>(ReadOnlySpan<T> x, ReadOnlySpan<T> y, Span<T> destination)
             where T : IBitwiseOperators<T, T, T>
         {
-            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsShort<T, BitwiseOrOperator<short>>(x, y, destination))
+            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsInt16<T, BitwiseOrOperator<short>>(x, y, destination))
             {
                 return;
             }
@@ -45,7 +45,7 @@ namespace System.Numerics.Tensors
         public static void BitwiseOr<T>(ReadOnlySpan<T> x, T y, Span<T> destination)
             where T : IBitwiseOperators<T, T, T>
         {
-            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsShort<T, BitwiseOrOperator<short>>(x, y, destination))
+            if (typeof(T) == typeof(Half) && TryBinaryBitwiseInvokeHalfAsInt16<T, BitwiseOrOperator<short>>(x, y, destination))
             {
                 return;
             }