This enables all of the following operators to be vectorized for T == Half:
Abs, Add, BitwiseAnd, BitwiseOr, Ceiling, CopySign, Cos, CosPi, Cosh, Decrement, DegreesToRadians, Divide, Exp, Exp10, Exp10M1, Exp2, Exp2M1, ExpM1, Floor, Hypot, Increment, Log, Log10, Log10P1, Log2, Log2P1, LogP1, Max, MaxMagnitude, MaxMagnitudeNumber, MaxNumber, Min, MinMagnitude, MinMagnitudeNumber, MinNumber, Multiply, Negate, OnesComplement, Reciprocal, Remainder, Round, Sigmoid, Sin, SinPi, Sinh, Sqrt, Subtract, Tan, TanPi, Tanh, Truncate, Xor.

I don't love that it required adding code to each of those methods, but I couldn't come up with anything better and kept the code per method to a minimum (I left the typeof(T) == typeof(Half) check at each call site, primarily for clarity).

I did not change the reductions, as anything that involved more than the same number of Half <-> float roundtrips as exists today ended up perturbing the results to the point where tests failed, and most of the reductions involve multiple operators.

Sample benchmarks:
```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Numerics.Tensors;

BenchmarkSwitcher.FromAssembly(typeof(Bench).Assembly).Run(args);

public class Bench
{
    private Half[] _x, _y, _d;

    [Params(1, 8, 800)]
    public int Length { get; set; }

    [GlobalSetup]
    public void Setup()
    {
        _x = new Half[Length];
        _y = new Half[Length];
        _d = new Half[Length];
        var random = new Random(42);
        for (int i = 0; i < Length; i++)
        {
            _x[i] = (Half)random.NextSingle();
            _y[i] = (Half)random.NextSingle();
        }
    }

    [Benchmark]
    public void Add() => TensorPrimitives.Add(_x, _y, _d);

    [Benchmark]
    public void Exp() => TensorPrimitives.Exp(_x, _d);

    [Benchmark]
    public void Floor() => TensorPrimitives.Floor(_x, _d);

    [Benchmark]
    public void Xor() => TensorPrimitives.Xor(_x, _y, _d);
}
```

On my machine with AVX256 (not AVX512)...

Before:

| Method | Length | Mean         |
|------- |------- |-------------:|
| Add    | 1      |    12.173 ns |
| Exp    | 1      |    13.374 ns |
| Floor  | 1      |     7.554 ns |
| Xor    | 1      |     6.763 ns |
| Add    | 8      |    55.226 ns |
| Exp    | 8      |    86.672 ns |
| Floor  | 8      |    45.207 ns |
| Xor    | 8      |     7.802 ns |
| Add    | 800    | 5,119.294 ns |
| Exp    | 800    | 8,246.779 ns |
| Floor  | 800    | 4,235.245 ns |
| Xor    | 800    |   340.764 ns |

After:
| Method | Length | Mean         |
|------- |------- |-------------:|
| Add    | 1      |    12.286 ns |
| Exp    | 1      |    21.634 ns |
| Floor  | 1      |     8.812 ns |
| Xor    | 1      |     6.544 ns |
| Add    | 8      |    23.209 ns |
| Exp    | 8      |    29.455 ns |
| Floor  | 8      |    17.650 ns |
| Xor    | 8      |     6.620 ns |
| Add    | 800    |   896.571 ns |
| Exp    | 800    | 1,629.144 ns |
| Floor  | 800    |   678.427 ns |
| Xor    | 800    |    26.408 ns |